Refactoring: modified Module to prepare introduction of Sandbox
The sandbox is now provided as a parameter of start, no longer in the
constructor of the module (to avoid circular dependencies for objects
creation).

I added checks for the initialization of the sandbox. Unless a sandbox
is provided in start(), all methods stop without doing anything
(because they cannot even log the issue, since the logger is part of the
sandbox).