Relaxed contraint on Core Module
I removed the check in addSubscriber that the same Subscriber has not
been add already. This check was of no use since the Subscriber instance
will typically be different, even when the event filter and the callback
within are identical. Thus there is no real use case behind and relaxing
the constraint simplifies the implementation and the tests.