Completed the Sandbox module, adding methods for local navigation
I added two methods getHash() and onHashChange(), aliases for the same
methods on lb.base.history, and a new method getLocation(), which
"sandboxes" window.location to avoid uncontrolled modification to the
current location.

I fixed a second bug in goog.History, with a duplicate firing of the
initial hash in FF (HTML5 mode).