Heavy refactoring in lb.base.history
The goal of this refactoring was to make this module take care of its
own initialization and destruction.

Initializing this module after page load has dreadful effects, resetting
the page to blank. To avoid this, the module now intializes during
loading of the script code, which will occur before page load for a
static script.

As a result, the default url used for the iframe in IE could not longer
be configured using lb.core.config or setConfig on lb.core.application.
The call would typically happen too late, after the initialization has
already occured. The default url is now read from the link with rel
'shortcut icon' in the head, if present. The default value is still
'/favicon.ico'.