Added Roadmap for v1.6.0
I am currently working on support of the internationalization planned for
version v1.5.0.

During this development, I noticed the need for application developers to add
custom methods in the Sandbox API, specific to their application. In the
context of internationalization, that means being able to define custom methods
for formatting of various types of values or parsing of user inputs.

I designed a simple plugin system for the Sandbox, which I plan to release in
the following version: v1.6.0.