Added comments related to API design in the initial description
This base API is still a moving target. I have now resumed working on the
design API, coming with a fresh mind after a span of a few months.

The new API will be both simpler and more elegant, avoiding some pitfalls in
current design.

First of all, the language field (getter/setter) selected for the application
will be removed, replaced with the direct use of 'lbLanguage' property stored
in lb.base.config.

Second, the language code will now be a mandatory parameter of getProperty,
since there is no default language anymore.

Third, getValueOf is removed. The underlying functionality is too abstract and
the desired extensibility will hopefully be provided by the plugin system for
the Sandbox API instead.

Also the getLanguageVariants() method will probably be removed as well,
replaced with getLanguageCodes() returning a list of all language codes which
have associated properties.

Last but not least, methods and parameters are renamed to disambiguate the use
of 'language'. The core concepts are now limited to 'language code' and
'language property' which come as sets in 'language properties'.