Refactoring: removed application language from lb.base.i18n
I removed the private field application language, as well as the getter and
setter. The getter method is actually still present, but returns the default
language configured in 'lbLanguage' property or the language of the browser.

This is part of a larger refactoring to reach a leaner base i18n API.
Eventually, there will no longer be a default language at this level, only at
the level of the Sandbox API, which will have to specify a language for lookup.