Removed getLanguageVariants() from lb.base.i18n
I replaced getLanguageVariants with calls to getProperty for unit tests of
addLanguageProperties, removed unit tests of getLanguageVariants, and removed
the function from the public API of lb.base.i18n.

The function is still used internally in lb.base.i18n. The next step will be to
inline and remove the function completely.