Added getBrowserLanguage() to base i18n API
I wanted to avoid the direct access to window.navigator from the Sandbox i18n
API, which lies at the Core level.

The principle is to restrict the dependency of each layer to the layer below,
avoiding in this case to take browser quirks into account in the Core.

  Core
---------
  Base
---------
 Browser

This is a simple method, that returns navigator.language when available (all
browsers but IE) and navigator.browserLanguage otherwise (IE). There is no
fallback in case none is available: a good fallback might be to return null,
here the method would return undefined instead. I could put one in place in
case there is ever an issue.