Added first unit tests for HTML templates
I noticed while developing these unit tests that it was painful to have to
include the filters parameter in every call while it was not actually used
in the code of the filtering function under test.

As a result, I dediced to make the parameter optional and move it from the
first position to the last position.