Restrict getLanguage() to element nodes in base i18n API
The move was motivated by the lack of support for ownerElement property in IE
up to and including IE8 in compatibility mode.

This prevents from offering a cross-browser version of getLanguage() for
attribute nodes. The method still works cross-browser for text nodes, and I
have preserved unit tests targetting text nodes for now.