Completed update of unit tests for core i18n API (language in lang attribute)
The config was still used in unit tests for get(), getString() and
filtetHtml(); I replaced its use with setting of the 'lang' attribute of
the root HTML element, mostly through the sandbox method selectLanguage().

Direct use of the 'lang' attribute remains in reset code:
document.documentElement.removeAttribute('')

Should this be part of the behavior of reset() in lb.base.i18n?