Bug fix: hyphen must be considered for inheritance of properties in getProperty
In previous implementation, case-insensitive substrings were all considered as
language codes with properties suitable for inheritance. But 'te' should not
inherit from 'test', only from 'te-*'. I added a unit test that fails and
updated the implementation to use lb.base.i18n.contains().

This revealed a bug in contains() itself: it seems that it is too strict and
does not consider equal language codes to be contained in each other...