Added call to setUp() at start of each unit test for Sandbox API
This restores a neutral state before calling each unit test. The only
processing currently is to remove the 'lang' attribute of the root document
element. This had an impact on the appearance of the page, with the font being
replaced when the test failed during filterHtml() and the lang remained set to
'te-ST'.

I also identified two more processings that should be moved to the setUp(),
which I annoted with TODO comments.