Added implementation of filterHtml() in Sandbox API
This completes the implentation of the i18n API of the Sandbox.

The implementation is based on applyFilters() from base templates module,
and filters defined in base HTML filters and base i18n filters:
- topDownParsing (HTML)
- replaceParams (HTML)
- filterByLanguage (i18n)
- setLanguage (i18n)

filterByLanguage() is actually a function that generates a filter function as
a closure wrapped around the language code to use for filtering.