Added support for internationalization (i18n)
This is a major version. I completed the support for i18n methods in the
Sandbox API:
- i18n.getLanguageList()
- i18n.getSelectedLanguage()
- i18n.selectLanguage()
- i18n.addLanguageProperties()
- i18n.get()
- i18n.getString()
- i18n.filterHtml()

Also, new in this version:
- added HTML index file in doc/ as entry point to the documentation
- enhancements and bug fixes in base modules for i18n and templates
- added hasAttribute() and constants in base dom module

In the next major version, I plan to add support for plugins to the Sandbox API
which should make it easier to extend and customize the API for the specific
needs of an application. I also intend to document the different extension
points of the library:
- configuration properties and custom factory
- custom modules with a custom sandbox
- Sandbox API plugins