Refactoring: move code of filterHtml from Sandbox to base i18n templates (3/3)
Last step of the refactoring: I reduced unit tests for filterHtml in the
Sandbox to 2 use cases, checking that the code works without entering into
dark corners. This reduces redundancy in unit tests while ensuring consistency.

I also plan to reduce slightly the documentation of filerHtml() in the base
i18n templates, removing explanations that may be superfluous outside of the
context of the Sandbox.