Updated doc of getString() and filterHtml() to reflect intendend changes
I will add a fallback mechanism for the lookup of properties for parameter
replacement, i.e. looking for missing properties in language properties.

I also removed a TODO comment in filterByHtml(). I intended to add an optional
parameter with the set of filters, which would default to the 4 actually
included. This does not seem relevant after all, since I do not feel that
sandboxed modules should be able to run any code of their choice as filters;
the extra argument would be intended only for developers of extensions to the
Sandbox API, who could just as easily call applyFilters directly.