Refactoring: moved implementation of getString() to base i18n data module
I also plan to merge the implementation of get() with getProperty() from the
base i18n data module.

The bigger plan is to add a fallback mechanism for getString() and filterHtml()
in case a property is not found in dynamic data object, to look up the value
in registered language properties.

I identified an additional refactoring needed in the base string template
module: the replacement function must be extracted and made as an optional
argument of replaceParams(). This way, getString() and filterHtml() can
customize the replacement while relying on the default behavior provided by
the extract function.