Updated expected documentation of base string template module
replaceParams will now return a the filter function. It will take as parameter
a function that returns the value for a parameter to replace. This will allow
to customize the replacement in internationalization functions, where I would
like to put a fallback mechanism in place to get the value from language
properties when not found in dynamic data.