Updated unit tests of replaceParams() in base string template module
The usage of the method is modified: replaceParams() now returns the filter to
be applied as a function. Instead of providing replaceParams directly to
applyFilters(), it must be called with a getter closure as argument to generate
the actual filter.

This change will allow to customize the replacement for i18n.