Refactoring: move getString() from base i18n data to base i18n templates
The move will avoid circular dependencies when getString() makes use of the
soon-to-be-added method withValuesFromDataOrLanguageProperties() located within
the base i18n template as well.

I also updated dependencies in unit tests and concat task in the build:
- base i18n / base i18n data no longer rely on base template / string templates

Thus base template modules are now grouped back together in lists of
dependencies.