Enhanced internationalization: fallback to language properties in templates
In getString() and filterHtml(), values for the replacement of parameters may
now come either from the given data object, or if not found, from language
properties in the scope of given/selected language.

This enhancement is the fruit of several refactoring of i18n APIs in the
base modules. There is no change to the public Sandbox API.