Added description for a bug fix version in v1.5.3 in ROADMAP
I noticed an issue in IE, using a custom factory: a failure during the
initialization of custom widgets caused an uncaught error which prevented
loading of following modules.

I will now strenghten the checks to ensure that any failures during the loading
of a module is caught, not only failures during the module start(). After all,
the framework (or here custom code which extends the framework) could fail as
well. This should not prevent other modules from loading. I guess that it will
be easier to spot issues when a single module is missing/unreactive than when
the faulty module and all following modules go AWOL.