Added description for enhancement release planned as v1.5.4 in ROADMAP
This enhancement is intended to handle better the following use case: during
internationalization, it may not be clear from the beginning how much
flexibility is needed for the translation of a text. I stumbled upon a case,
where a parameter, which was initially intended to contain a date, would then
only contain the time when the date is today. This had an impact on the
translation, where the prefix 'on' was fixed, and now needed to be 'on' for
date values and 'at' for time values. This kind of logic would depend on the
language, and would better be left to the i18n properties. Therefore, I intend
to add support for functions for value replacement in getString() and
filterHtml(); when a function is found for the replacement of a value, it will
be called with the data object as argument. The function can now build the
corresponding message with a custom logic which depends on the language.

Note that the global code may still need to be modified, to add more context
as needed by the localization function, but at least the logic which depends
on the language can be handled in the file where language-specific properties
are defined.