Enhancement: added explicit string conversion for Module
I declared toString() as an alias for getId(), which enhances log messages
printed for the failure of a module during startAll() or endAll().

This mechanism is flexible enough that if the failure is due to the module
being null undefined or the wrong kind of object, the native toString()
function is used instead.