Added unit tests for topDownParsing(), failing with current version
I wish to strengthen the HTML filters and report errors which occur in filters.
I have spent quite some time today tracking a bug in IE where a filter fails
on an attribute of an element deep down in the tree, and IE does not break on
the error. Having a try/catch close to the error, reporting the name of the
attribute and the type, id, className if available of the parent node would
certainly help a lot in this kind of situation.