Added try/catch with log message when applying filters in topDownParsing
A different log is printed for attributes and child nodes. I tried to provide
enough context to localize the error quickly while avoiding to provide too
much details which could be counter-productive.

Still, the log messages are quite long and may reduce the readability of the
function. There is also some redundancy between the log message for attributes
and the log message for child nodes. This may benefit from a refactoring,
extracting logging to a separate function.