Enhancement: log errors from filters in topDownParsing()
These logs help to catch errors when a filter fails for an unknown reason,
e.g. for the type attribute of input elements in IE, which can only be written
once, before the element is part of the DOM...

In order to reduce this kind of error, I will modify the behavior of the
replaceParams() filter in base HTML templates module, to set the new nodeValue
only when it differs from the previous nodeValue.