Added description for next bug fix version in ROADMAP, for an issue in IE
I also added a TODO comment in replaceParams(), where the change is indented.

The input.type attribute is readonly when the element is in the DOM. I did not
find any way to discover that setting the value of this kind of attribute will
fail in the general case, and thus decided to avoid setting the value unless
a replacement actually occured (in which case the developer may have a way to
figure out what happens when it fails, if only because he will not find the
updated value corresponding to a parameter that he specified).