Bug fix: Fixed replacement of parameters in src and href in IE7
In IE7, the base location of the page was left at the start of the src and href
attributes, after replacement.

For example, a parameter #url# in a link href ended up as:
  http://legal-box.net/webmail/http://legal-box.com/legal
instead of
  http://legal-box.com

I applied a known fix which I had developed for the bezen.org JavaScript
library: the base location of the page is removed from the start of the URL
before replacement. The fix is applied both to src and href attributes.