Parameter replacement in getString() is now recursive
Also, getString() is now used to replace parameters recursively in HTML
templates for internationalization.

I added an extension point in getString(): a function can be set as the value
of a parameter in language properties. This allows to customize a value, by
computing it based on param name, data for context, and language code, instead
of providing a fixed string value.