Added lb.base.type module, with first implementation of is()
is() is a utility method to check the type of a value based on its internal
class.

This first implementation is working, but should be refined:
  - adding unit tests for string types 'null' and 'undefined'
  - adding unit tests for 'regexp' type for regular expressions

The documentation shall be completed with a list of expected types for various
examples and the details of the limitations, values whose type varies depending
on the browser.