Updated behavior of has(), now accepting inherited properties as well
Restricting has() to own properties happens to cause more harm than it was
supposed to prevent: when working with BOM and DOM objects, many properties
are inherited, which is not an issue as long as they are read, not written to.

has() is now a better complement to is(), which covers more ground while is()
is more strict. For example, in lb.base.dom.hasAttribute(), I started with a
check using is:

    is(element,'hasAttribute','function')

which failed in IE where hasAttribute is an object, but can still be called.
I then replaced it with has:

    has(element,'hasAttribute')

which now works with this change, but did not before (hasAttribute is inherited
in Firefox and probably other browsers as well).