Added has() and is() methods to Sandbox utils API
These two utilities are also used internally. has() checks for the existence of
a property, with a non-null, non-undefined value. is() does the same, and
verifies the type of the value as well. These two methods avoid relying on
type coercion or using the typeof operator, which is broken.

is() is based on the internal class of values, which reflects more accurately
the type of a value (e.g. 'array' instead of 'object' for arrays).