Enhancement: speed up build using jslint4java
I added jslint4java by Dominic Mitchell, and defined an updated macro to check
either a single file or all JavaScript files in a directory.

The old macro used to fork the VM and spawn a different process for each file
to check, which was very long. The new macro is much faster, which reduces the
whole build to about 30s.