Remove xml:lang on html element of page for all unit tests, for IE9's sake
This is the same change I did in previous commit for the unit tests of
lb.base.i18n.js. It happens that calling removeAttribute() is not enough to
remove the lang attribute in IE9 - it subsists if an xml:lang has been defined.

Omitting the xml:lang attribute in the first place solves the issue.

I tried also to remove the xml:lang attribute programmatically using
removeAttribute('xml:lang') and
removeAttributeNS('http://www.w3.org/XML/1998/namespace','lang'), but both
resulted in a page crash in IE9.