Commented out conditional code that will never run
Conditional blocks with dead code are optimized out by the Google Closure
Compiler. This is a hand-made optimization, commenting out large chunks of
code that will never run because we are not using advanced functionalities and
corresponding flags are disabled.

There is one flag still enabled, which could be reconsidered (I do not fully
understand the impact yet):
    goog.debug.LogRecord.ENABLE_SEQUENCE_NUMBERS