Bug fix: added property for parent path to allow relative resolution of paths
The basedir property is overridden in the main build script, and thus cannot
be used in the imported script to find the relative location of libraries.

I followed the explanations from import task documentation in Ant manual [1]
to define a property with the location of the parent folder of macros.xml.

[1] Import - Ant Manual
http://ant.apache.org/manual/Tasks/import.html