Use project name as a prefix in the name of generated files
I am now using 'lb' as the name of the project, instead of 'lb_js_scalableApp',
to use the project name as a prefix in the name of generated files.

This will allow to import this script in different projects, and generated
files will start with a different prefix, configured as the name of each
project.

The only target that needs to be redefined in each project is "combine",
which lists scripts to combine for the output.