Enhancement: cache busting timestamp now includes time
The timestamp included in minified script for production to force browsers
to download the latest version of the script used to include only the date.
This could be problematic when several versions were released on the same day.

The timestamp now includes the time as well, in hours and minutes. Which should
be fine as long as there are never multiple releases in the same minute.