Replaced the token modules.list with modules.array
I also removed the start/end of the array from the template file; a full
JSON array must now be provided in the replacement value.

I did not like having to document this malformed JSON array format. I later
figured out that I could create an intermediate property in Ant script to
add the start and end characters wrapping the list to make it a proper array.