Fixed location of output directory for requireJS, fixed location of build.js
An absolute location is needed for the output directory, otherwise the location
is relative to the path of the build config file.

I moved the build.js file to build/build.js, because it looks for require.js
in the parent directory by default. This is needed for use with includeRequire
option.