Check modules src or src/{project name} for compatibility when imported
When the ant script is imported for use in other modules, src/{project name}
may not be available. If not, check the whole src instead.

I have to handle src/ with care, because other libraries, such as closure,
may break the build by failing JSLint checks. Only our libraries should break
the build, not external ones.

I also added a target to check amd with JSLint.