Patch for requireJS parse.js, from trunk of jrburke/r.js
This patch ensures that the module "amd/define" is included first in the
combined file, by looking for the assignment "define.amd" which corresponds
to a claim that the library supports CommonJS AMD.

Without the patch, "amd/define" happens to be put in the first position, just
because it is listed first in alphabetical order.

The change occurs in isRequireNode/isDefineNode in parse.js, which used to
check for the presence of an assignment to non-standard property "require.s".