Enhancement: fixed and clarified documentation
I removed a note about the initial design of the core application, which is no
longer relevant: there is now a method to unsubscribe() in the Sandbox API.

I fixed the diagram for the actual implementation of Core Application: the
methods setElementFactory() and getElementFactory() have been removed,
replaced with setOptions() which allows to configure both the element factory
and the plugins builder.

I added curly braces around "template" names in diagrams,
e.g. "lb.ui.{moduleName}".