Added amd/define module at start of bezen.org/bezen.js for compatibility
In case define() is not defined, amd/define is used.

This script is included in bezen.js for compatibility with projects that
depend on the scripts in bezen.org for unit tests.

It may be removed in the future when all these projects have been converted
to use requireJS module definition.